tests/wpt/tests/css/css-syntax/charset/page-windows-1251-css-http-bogus.html

<!doctype html>
<title>CSS charset: page windows-1251, CSS HTTP bogus</title>
<link rel=help href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding">
<meta charset=windows-1251>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel=stylesheet href="support/http-bogus.bogus.css">
<div id=log></div>
<div id=&#x0418;></div>
<script>
var elm = document.getElementById('\u0418');
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm, '').visibility, 'hidden');
  this.done();
});
</script>
